requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30
sqlalchemy[asyncio]>=2.0,<3.0
asyncpg>=0.29
alembic>=1.13
pydantic-settings>=2.0
pyjwt>=2.8
bcrypt>=4.0
python-multipart>=0.0.9
pillow>=10.0
python-slugify>=8.0

# Dev / testing
pytest>=8.0
pytest-asyncio>=0.23
httpx>=0.27
psycopg2-binary>=2.9
